+++ src/Completions/Sources/RazorCompletionSource.cs
@@ -166,8 +166,67 @@ namespace TailwindCSSIntellisense.Completions.Sources
             }
             else
             {
-                classText = null;
-                return false;
+                var segments = text.Substring(quotationMarkAfterLastClassAttribute + 1).Split([' '], StringSplitOptions.RemoveEmptyEntries);
+
+                bool isInRazor = false;
+                int depth = 0;
+                // Number of quotes (excluding \")
+                // Odd if in string context, even if not
+                int numberOfQuotes = 0;
+
+                foreach (var segment in segments)
+                {
+                    if (segment.StartsWith("@") || isInRazor)
+                    {
+                        bool isEscaping = false;
+
+                        foreach (var character in segment)
+                        {
+                            bool escape = isEscaping;
+                            isEscaping = false;
+
+                            if (numberOfQuotes % 2 == 1)
+                            {
+                                if (character == '\\')
+                                {
+                                    isEscaping = true;
+                                }
+                            }
+                            else
+                            {
+                                if (character == '(')
+                                {
+                                    depth++;
+                                }
+                                else if (character == ')')
+                                {
+                                    depth--;
+                                }
+                            }
+
+                            if (character == '"' && !escape)
+                            {
+                                numberOfQuotes++;
+                            }
+                        }
+
+                        isInRazor = depth != 0 || numberOfQuotes % 2 == 1;
+                    }
+                    else if (segment.Contains('"'))
+                    {
+                        classText = null;
+                        return false;
+                    }
+                }
+
+                if (depth != 0 || numberOfQuotes % 2 == 1)
+                {
+                    classText = null;
+                    return false;
+                }
+
+                classText = text.Substring(quotationMarkAfterLastClassAttribute + 1);
+                return true;
             }
         }