+++ src/QuickInfo/HtmlQuickInfoSource.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.Language.Intellisense;
+﻿using EnvDTE80;
+using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Language.StandardClassification;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Adornments;
@@ -44,7 +45,7 @@ namespace TailwindCSSIntellisense.QuickInfo
             if (lastQuotationMark == quotationMarkAfterLastClassAttribute)
             {
                 var startIndex = lastQuotationMark + 1;
-                text = text.Substring(lastQuotationMark + 1);
+                text = text.Substring(startIndex);
                 startIndex += text.LastIndexOf(' ') == -1 ? 0 : text.LastIndexOf(' ') + 1;
                 var length = 1;
 
@@ -66,8 +67,85 @@ namespace TailwindCSSIntellisense.QuickInfo
             }
             else
             {
-                span = null;
-                return false;
+                var end = searchPos.Value;
+
+                bool isInRazor = false;
+                int depth = 0;
+                // Number of quotes (excluding \")
+                // Odd if in string context, even if not
+                int numberOfQuotes = 0;
+                bool isEscaping = false;
+
+                char[] endings = ['"', '\''];
+
+                while (end < _textBuffer.CurrentSnapshot.Length - 1 && (depth != 0 || numberOfQuotes % 2 == 1 || endings.Contains(end.GetChar()) == false))
+                {
+                    var character = end.GetChar();
+
+                    if (character == '@')
+                    {
+                        isInRazor = true;
+                    }
+                    else if (isInRazor)
+                    {
+                        if (searchPos == end)
+                        {
+                            span = null;
+                            return false;
+                        }
+
+                        bool escape = isEscaping;
+                        isEscaping = false;
+
+                        if (numberOfQuotes % 2 == 1)
+                        {
+                            if (character == '\\')
+                            {
+                                isEscaping = true;
+                            }
+                        }
+                        else
+                        {
+                            if (character == '(')
+                            {
+                                depth++;
+                            }
+                            else if (character == ')')
+                            {
+                                depth--;
+                            }
+                        }
+
+                        if (character == '"' && !escape)
+                        {
+                            numberOfQuotes++;
+                        }
+
+                        if (depth == 0 && numberOfQuotes % 2 == 0 && character == ' ')
+                        {
+                            isInRazor = false;
+                        }
+                    }
+                    else if (character == ' ' && searchPos <= end)
+                    {
+                        break;
+                    }
+
+                    end += 1;
+                }
+
+                if (depth != 0 || numberOfQuotes % 2 == 1 || searchPos > end)
+                {
+                    span = null;
+                    return false;
+                }
+
+                var startIndex = lastQuotationMark + 1;
+                text = text.Substring(startIndex);
+                startIndex += text.LastIndexOf(' ') + 1;
+
+                span = new SnapshotSpan(new SnapshotPoint(_textBuffer.CurrentSnapshot, startIndex), end);
+                return true;
             }
         }
     }