+++ src/source.extension.cs
@@ -9,9 +9,9 @@ namespace TailwindCSSIntellisense
     {
         public const string Id = "TailwindCSSIntellisense.b29696d4-76a4-43d1-a203-c8e692368847";
         public const string Name = "Tailwind CSS VS2022 Editor Support";
-        public const string Description = @"Unofficial VS2022 extension for Tailwind CSS, offering editor features such as IntelliSense, build integration, and more.";
+        public const string Description = @"Unofficial VS2022 extension for Tailwind CSS, offering editor features such as IntelliSense, linting, build shortcuts, and more.";
         public const string Language = "en-US";
-        public const string Version = "1.1.9.3";
+        public const string Version = "1.2.2";
         public const string Author = "Theron Wang";
         public const string Tags = "tailwind, tailwind css, css, html, cshtml, razor, blazor, intellisense, auto-completion";
     }